Update variables.tf
Allow `*` in repo names.

In cases where there are a lot of repositories that make up the AWS policy, we hit the "Role trust policy length". By default this limit is 2048 chars, expandable to 4096 as a hard limit.

This "fix" allows wildcard char `*` to be added to similar repo names.